backend/requirements.txt
# 🤖🪞 Mirror Mirror Backend Requirements
# "Because even fake therapy needs real dependencies"

# Core Flask Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
python-socketio==5.9.0

# File Upload & Processing
Werkzeug==2.3.7
Pillow==10.0.1
opencv-python==4.8.1.78

# AI & Machine Learning
openai==0.28.1
transformers==4.35.2
torch>=2.2.0
torchvision>=0.17.0
face-recognition==1.3.0
mediapipe==0.10.8

# Image/Video Processing
moviepy==1.0.3
imageio==2.31.5
numpy==1.24.3
scipy==1.11.4

# HTTP Requests & APIs
requests==2.31.0
python-dotenv==1.0.0

# Authentication & Security
PyJWT==2.8.0
bcrypt==4.1.2

# Background Task Processing
celery==5.3.4
redis==5.0.1

# Utilities
python-dateutil==2.8.2
uuid==1.30
